+++ Arma3ObjectBuilder/tool_materials/colors.py
