+++ Arma3ObjectBuilder/ui/__init__.py
@@ -1,48 +1,3 @@
-if "props_object_mesh" in locals():
-    from importlib import reload
-
-    if "import_export_armature" in locals():
-        reload(import_export_armature)
-    if "import_export_asc" in locals():
-        reload(import_export_asc)
-    if "import_export_mcfg" in locals():
-        reload(import_export_mcfg)
-    if "import_export_p3d" in locals():
-        reload(import_export_p3d)
-    if "import_export_rtm" in locals():
-        reload(import_export_rtm)
-    if "import_export_tbcsv" in locals():
-        reload(import_export_tbcsv)
-    if "import_export_paa" in locals():
-        reload(import_export_paa)
-    if "props_action" in locals():
-        reload(props_action)
-    if "props_material" in locals():
-        reload(props_material)
-    if "props_object_mesh" in locals():
-        reload(props_object_mesh)
-    if "tool_outliner" in locals():
-        reload(tool_outliner)
-    if "tool_hitpoint" in locals():
-        reload(tool_hitpoint)
-    if "tool_mass" in locals():
-        reload(tool_mass)
-    if "tool_materials" in locals():
-        reload(tool_materials)
-    if "tool_paths" in locals():
-        reload(tool_paths)
-    if "tool_proxies" in locals():
-        reload(tool_proxies)
-    if "tool_rigging" in locals():
-        reload(tool_rigging)
-    if "tool_scripts" in locals():
-        reload(tool_scripts)
-    if "tool_utilities" in locals():
-        reload(tool_utilities)
-    if "tool_validation" in locals():
-        reload(tool_validation)
-
-
 from . import import_export_armature
 from . import import_export_asc
 from . import import_export_mcfg
@@ -56,7 +11,6 @@ from . import props_object_mesh
 from . import tool_outliner
 from . import tool_hitpoint
 from . import tool_mass
-from . import tool_materials
 from . import tool_paths
 from . import tool_proxies
 from . import tool_rigging
